Factor out socket pipe magic to separate function
This patch breaks out the the main() receive/send loop into a separate
function.  With the intention to make it easier to understand.  For the
first shot however the main focus was to get inetd mode working, so the
code is slightly bloated.

Also, change all printf()'s to use syslog() instead.  Pretty useless to
rely on stdout/stderr when running in inetd mode.  Also added a few
debug messages which currently default to on.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>