Linger in inetd mode for a while in case of more requests
This reduces the CPU load slightly when processing multiple requests to
the same UDP port during a short time frame, e.g. not having to fork and
start a new uredir for every OID during an SNMP walk.  Three seconds
should be enough for everybody, if not we can add a -t SEC configurable
timeout later.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>