Buffersize reduced over time
The buffer used to forward data is used in both directions and when
forward data from inner (conn) to outer the specified size of the
buffer was altered making the specified size smaller than the actual
size of the buffer.

When the connection is reused the buffer sizes in the header is now
reset.